Устраняет замечания по 4-у ДЗ и отправляет 5-е ДЗ
