+++ js/consts.js
@@ -1,4 +1,4 @@
-export const DESCRIPTION = [
+const DESCRIPTION = [
   'Закат',
   'Цветы',
   'Спортивный атомобиль',
@@ -8,13 +8,23 @@ export const DESCRIPTION = [
   'кот Кекс',
 ];
 
-export const MESSAGES = [
+const MESSAGES = [
   'Всё отлично!',
-  'В целом всё неплохо. Но не всё.'
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
 ];
 
-export const NAMES = [
+const NAMES = [
   'Виктор',
   'Иван',
   'Мария',
+  'Светлана',
+  'Сергей',
+  'Дмитрий',
+  'Наталья',
 ];
+
+export {DESCRIPTION, MESSAGES, NAMES};
